Snow Halation, first break and second verse
Exactly 5:00 into this song. Frankly I'm getting kind of scared of how huge this is becoming.
It took four days to write the last 12 bars. I'm not even very fond of its showiness (not nearly as much as RayHu92 I'll bet) but couldn't think of anything better.
It would be amazing if I could finish this for Christmas.