Snow Halation, sketches for second half
Also some modifications for the first half.

Going to try a different plan for the second half, in an attempt to write faster.
The "second half" file currently consists of rough sketches of what patterns to use for what parts,
in the next approx. three minutes of the song.
This file will be gradually updated and filled out over the next few weeks.
The last two minutes will be similar to the original verse, chorus, and finale arrangements.